feat: refactor React plugin management and add plugin setup functionality
